break out PartyList component, set to root path
